AAF-4 PAgination working and all tests

+++ services/NotificationService.js
@@ -2,7 +2,7 @@ const MongooseService = require("./MongooseService.js");
 const model = require("../database").getModel("notification");
 const httpError = require("http-errors");
 // Check if the ID given is a valid MongoDB ObjectID
-const isIdValid = require("./utilities").isIdValid;
+const isIdValid = require("../middleware/validation/utilities").isIdValid;
 
 class NotificationService {
     /**
@@ -25,6 +25,7 @@ class NotificationService {
         if (!validateNotification(notificationToCreate)) {
             throw httpError(400, "Notification data is invalid.");
         }
+
         const notification = {
             message: notificationToCreate.message,
             username: notificationToCreate.username,
@@ -54,14 +55,24 @@ class NotificationService {
      * @returns {httpError} 200 If deleteing the Notification is successful.
      * @returns {httpError} 404 If notification could not be deleted.
      */
-    async deleteNotification(notificationToDelete) {
+    async deleteNotification(notificationToDelete, requestedBy) {
         if (!isIdValid(notificationToDelete)) {
             throw httpError(404, "Notification ID is invalid.");
         }
         return this.mongooseService
-            .deleteById(notificationToDelete)
-            .catch((error) => {
-                throw httpError(404, error.message);
+            .findById(notificationToDelete)
+            .then((notification) => {
+                if (!validateNotificationPerm(requestedBy, notification)) {
+                    throw httpError(
+                        403,
+                        "You do not have permission to delete this notification."
+                    );
+                }
+                return this.mongooseService
+                    .deleteById(notificationToDelete)
+                    .catch((error) => {
+                        throw httpError(404, error.message);
+                    });
             });
     }
 }
@@ -78,5 +89,15 @@ function validateNotification(notification) {
         return true;
     }
 }
-
+function validateNotificationPerm(requestedBy, notification) {
+    if (
+        requestedBy.role !== "Admin" &&
+        requestedBy.username.toLowerCase() !==
+            notification.username.toLowerCase()
+    ) {
+        return false;
+    } else {
+        return true;
+    }
+}
 module.exports = NotificationService;
